Lucky Duck Casino/give/requirements.txt
uvicorn
fastapi
sqlalchemy
databases
werkzeug
pyjwt
pycryptodome
pydantic
pydantic-settings
sqlmodel
psycopg2-binary 
jinja2